requirements.txt
wheel
requests==2.24.0
pytest==5.3.5
PyHamcrest==2.0.2
pytest-allure-dsl>=0.0.6,<0.1
jsonschema==3.2.0
eventlet==0.28.0
